Move icon components to icons folder
