+++ components/icons/paint-icon.js
