+++ components/icons/star.js
